Tracing: expose directory path setting as tracing.dir
in rabbitmq.conf.

Note that this does not include any tests because
the test would have to use a writeable directory,
and it is not obvious what kind of cross-platform
path that is not computed programmatically they
could use.

It's a trivial schema file that uses an existing
core validator => let's leave it as is.