CI: Turn off mixed version tests running Khepri
Khepri is experimental and not guaranteed to be upgradable. We should
re-enable mixed version testing as Khepri stabilizes but for now we
should turn it off.

This saves some CI time and avoids some failures that will be introduced
when we upgrade Khepri to v0.14.0.

(cherry picked from commit e08310ab131012d6442c53b91ef60ec611885e36)