rabbitmq-upgrade revive: handle more errors
returned by Ra, e.g. when a replica cannot be
restarted because of a concurrent delete
or because a QQ was inserted into a schema data
store but not yet registered as a process on
the node.

References #12013.

(cherry picked from commit 2f165e02f2e57a3a2f1bf379579f6ff4d46cf324)