Use SASL layer in AMQP 1.0 Erlang client
This commit ensures that AMQP 1.0 shovels will always use a
SASL security layer when connecting to an AMQP 1.0 broker.
Instead of skipping SASL, the client will use SASL mechanism ANONYMOUS.

https://github.com/rabbitmq/rabbitmq-server/pull/11999 mandates that
AMQP 1.0 clients use a SASL security layer when connecting to a RabbitMQ
4.0 node.

This commit is only applied to the `v3.13.x` branch such that a shovel
running on a 3.13.7 node will be able to connect via AMQP 1.0 to a
RabbitMQ 4.x node.