Khepri: Filter running nodes when selecting a node to cluster with
[Why]
So far, the code that selected the node to use as the "entry point" to
add the local node to a remote cluster assumed that all cluster members
were running and it picked the first node in the cluster members list.

If that node was stopped, the join would fail immediately, even if the
rest of the members were running fine.

[How]
Now the function filters out nodes that are unavailable or don't run the
expected Khepri store. Then it uses the resulting list as before.

The code returns an error if all nodes are stopped or unreachable.