Add connection duration to "closing AMQP connection..." msgs
A while back, @mkuratczyk noted that we keep the timestamp of when a
connection is established in the connection state and related ETS table.

This PR uses the `connected_at` timestamp to calculate the duration of
the connection, to make it easier to identify short-running connections
via the log files.

(cherry picked from commit 3440e374d06295cddbf0b49a5d6bb1386b086d0f)