Tweak wording around experimental feature flags (Khepri)
This updates Khepri FF description to be more correct
and to the point.

It also tweaks the management UI copywriting so
that it does not recommend against the use of
Khepri in production as it is much more mature
in 4.0.

(cherry picked from commit 29fcb33a6b8f09ade5c55d8f91d0712f99f80d3c)